refactor: split connection node from common for later endpoint
Closes #99
