+++ autopush-common/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "autopush_common"
+version = "1.0.0"
+authors = [
+  "Ben Bangert <ben@groovie.org>",
+  "JR Conlin <jrconlin@mozilla.com>",
+  "Alex Crichton <alex@alexcrichton.com>",
+  "Phil Jenvey <pjenvey@underboss.org>",
+]
+edition = "2018"
+
+[lib]
+name = "autopush_common"
+
+[dependencies]
+cadence = "0.16.0"
+chrono = "0.4.6"
+config = "0.9.2"
+error-chain = "0.12.0"
+futures = "0.1.25"
+futures-backoff = "0.1.0"
+httparse = "1.3.3"
+# XXX: pin to hyper 0.11 for now: 0.12 has many changes..
+hyper = "0.11.27"
+lazy_static = "1.2.0"
+mozsvc-common = "0.1.0"
+rand = "0.6.3"
+regex = "1.1.0"
+reqwest = "0.9.5"
+rusoto_core = "0.36.0"
+rusoto_credential = "0.15.0"
+rusoto_dynamodb = "0.36.0"
+sentry = { version = "0.13.0", features = ["with_error_chain"] }
+serde = "1.0.84"
+serde_derive = "1.0.84"
+serde_dynamodb = "0.2.1"
+serde_json = "1.0.34"
+slog = { version = "2.4.1", features = ["max_level_trace", "release_max_level_info"] }
+slog-async = "2.3.0"
+slog-term = "2.4.0"
+slog-mozlog-json = "0.1.0"
+slog-scope = "4.1.1"
+slog-stdlog = "3.0.2"
+tokio-core = "0.1.17"
+tungstenite = { version = "0.6.1", default-features = false }
+uuid = { version = "0.7.1", features = ["serde", "v4"] }
