+++ autopush-common/src/errors.rs
