+++ autopush/Cargo.toml
@@ -1,0 +1,57 @@
+[package]
+name = "autopush"
+version = "1.53.1"
+authors = [
+  "Ben Bangert <ben@groovie.org>",
+  "JR Conlin <jrconlin@mozilla.com>",
+  "Alex Crichton <alex@alexcrichton.com>",
+  "Phil Jenvey <pjenvey@underboss.org>",
+]
+edition = "2018"
+
+[[bin]]
+name = "autopush_rs"
+path = "src/main.rs"
+
+[dependencies]
+autopush_common = { path = "../autopush-common" }
+base64 = "0.10.0"
+bytes = "0.4.11"
+cadence = "0.16.0"
+chan-signal = "0.3.2"
+config = "0.9.2"
+docopt = "1.0.2"
+error-chain = "0.12.0"
+fernet = "0.1.0"
+futures = "0.1.25"
+hex = "0.3.2"
+httparse = "1.3.3"
+# XXX: pin to hyper 0.11 for now: 0.12 has many changes..
+hyper = "0.11.27"
+lazy_static = "1.2.0"
+mozsvc-common = "0.1.0"
+openssl = "0.10.16"
+reqwest = "0.9.5"
+rusoto_dynamodb = "0.36.0"
+sentry = { version = "0.13.0", features = ["with_error_chain"] }
+serde = "1.0.84"
+serde_derive = "1.0.84"
+serde_dynamodb = "0.2.1"
+serde_json = "1.0.34"
+slog = { version = "2.4.1", features = ["max_level_trace", "release_max_level_info"] }
+slog-async = "2.3.0"
+slog-term = "2.4.0"
+slog-mozlog-json = "0.1.0"
+slog-scope = "4.1.1"
+# state_machine_future = { version = "0.1.6", features = ["debug_code_generation"] }
+state_machine_future = "0.2.0"
+time = "0.1.41"
+tokio-core = "0.1.17"
+tokio-io = "0.1.10"
+tokio-openssl = "0.3.0"
+tokio-service = "0.1.0"
+tokio-tungstenite = { version = "0.6.0", default-features = false }
+tungstenite = { version = "0.6.1", default-features = false }
+uuid = { version = "0.7.1", features = ["serde", "v4"] }
+# XXX: pin woothee until >= 0.8.1
+woothee = "0.7.3"
