+++ autopush/src/server/protocol.rs
@@ -1,4 +1,4 @@
-//! Definition of Internal Router, Python, and Websocket protocol messages
+//! Definition of Internal Router and Websocket protocol messages
 //!
 //! This module is a structured definition of several protocol. Both
 //! messages received from the client and messages sent from the server are
@@ -13,7 +13,14 @@ use serde_derive::{Deserialize, Serialize};
 use serde_json;
 use uuid::Uuid;
 
-use crate::util::ms_since_epoch;
+use autopush_common::notification::Notification;
+
+#[derive(Serialize)]
+#[serde(untagged)]
+pub enum BroadcastValue {
+    Value(String),
+    Nested(HashMap<String, BroadcastValue>),
+}
 
 // Used for the server to flag a webpush client to deliver a Notification or Check storage
 pub enum ServerNotification {
@@ -88,13 +95,6 @@ pub struct ClientAck {
 }
 
 #[derive(Serialize)]
-#[serde(untagged)]
-pub enum BroadcastValue {
-    Value(String),
-    Nested(HashMap<String, BroadcastValue>),
-}
-
-#[derive(Serialize)]
 #[serde(tag = "messageType", rename_all = "snake_case")]
 pub enum ServerMessage {
     Hello {
@@ -137,63 +137,4 @@ impl ServerMessage {
             _ => serde_json::to_string(self),
         }
     }
-}
-
-#[derive(Serialize, Default, Deserialize, Clone, Debug)]
-pub struct Notification {
-    #[serde(rename = "channelID")]
-    pub channel_id: Uuid,
-    pub version: String,
-    #[serde(default = "default_ttl", skip_serializing)]
-    pub ttl: u64,
-    #[serde(skip_serializing)]
-    pub topic: Option<String>,
-    #[serde(skip_serializing)]
-    pub timestamp: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<String>,
-    #[serde(skip_serializing)]
-    pub sortkey_timestamp: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub headers: Option<HashMap<String, String>>,
-}
-
-impl Notification {
-    /// Return an appropriate sort_key to use for the chidmessageid
-    ///
-    /// For new messages:
-    ///     02:{sortkey_timestamp}:{chid}
-    ///
-    /// For topic messages:
-    ///     01:{chid}:{topic}
-    ///
-    /// Old format for non-topic messages that is no longer returned:
-    ///     {chid}:{message_id}
-    pub fn sort_key(&self) -> String {
-        let chid = self.channel_id.to_hyphenated();
-        if let Some(ref topic) = self.topic {
-            format!("01:{}:{}", chid, topic)
-        } else if let Some(sortkey_timestamp) = self.sortkey_timestamp {
-            format!(
-                "02:{}:{}",
-                if sortkey_timestamp == 0 {
-                    ms_since_epoch()
-                } else {
-                    sortkey_timestamp
-                },
-                chid
-            )
-        } else {
-            // Legacy messages which we should never get anymore
-            format!("{}:{}", chid, self.version)
-        }
-    }
-
-    pub fn expired(&self, at_sec: u64) -> bool {
-        at_sec >= self.timestamp as u64 + self.ttl as u64
-    }
-}
-
-fn default_ttl() -> u64 {
-    0
 }
